% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.fuzzify_vector}
\alias{.fuzzify_vector}
\title{Wewnętrzna transformacja rozmyta (TFN)}
\usage{
.fuzzify_vector(vector_values)
}
\description{
Przekształca wartość ostrą (crisp) na
Trójkątną Liczbę Rozmytą (TFN) w postaci (l, m, u).
}
\keyword{internal}
