% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.parse_ide_model}
\alias{.parse_ide_model}
\title{Wewnętrzny parser modelu decyzyjnego IDE}
\usage{
.parse_ide_model(model_syntax)
}
\description{
Funkcja pomocnicza interpretująca składnię definicji
obszarów decyzyjnych w pakiecie IDEChoiceR.
Przekształca zapis tekstowy typu:
"Wydajnosc =~ start + ram" na strukturę listową w R.
}
\keyword{internal}
