% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.scale_to_mcda}
\alias{.scale_to_mcda}
\title{Wewnętrzna normalizacja do skali 1–9}
\usage{
.scale_to_mcda(vector_values)
}
\description{
Skaluje dowolne wartości liczbowe do przedziału 1–9
zgodnie z wymaganiami metod MCDA (TOPSIS, VIKOR).
}
\keyword{internal}
