% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{idechoice_sample_data}
\alias{idechoice_sample_data}
\title{Przykładowe dane do analizy wyboru środowiska IDE}
\format{
Ramka danych (data frame) z 20 wierszami i 14 zmiennymi:
\describe{
\item{EkspertID}{Identyfikator osoby oceniającej lub panelu eksperckiego}
\item{IDE}{Nazwa ocenianego środowiska programistycznego}
\item{wydajnosc_start}{Czas uruchamiania IDE (w sekundach)}
\item{wydajnosc_zuzycieRAM}{Średnie zużycie pamięci RAM (MB)}
\item{funkcjonalnosc_debugger}{Ocena debuggera (skala 1–5)}
\item{funkcjonalnosc_pluginy}{Liczba dostępnych rozszerzeń / wtyczek}
\item{funkcjonalnosc_git}{Ocena integracji z systemem kontroli wersji (1–5)}
\item{uzytecznosc_interfejs}{Ocena intuicyjności interfejsu (1–5)}
\item{uzytecznosc_personalizacja}{Możliwość personalizacji (1–5)}
\item{uzytecznosc_dokumentacja}{Jakość dokumentacji i wsparcia (1–5)}
\item{koszt_licencja}{Koszt licencji (w USD lub PLN)}
\item{koszt_subskrypcja}{Koszt miesięcznej subskrypcji}
\item{stabilnosc_bledy}{Średnia liczba błędów krytycznych w miesiącu}
\item{popularnosc_spolecznosc}{Aktywność społeczności (np. liczba wątków / wsparcie)}
}
}
\usage{
data(idechoice_sample_data)
}
\description{
Zbiór danych zawierający przykładową macierz decyzyjną dla 20 środowisk
programistycznych (IDE) ocenianych w ramach 13 zmiennych
reprezentujących 4 główne obszary decyzyjne
(Wydajność, Funkcjonalność, Użyteczność, Koszt).
}
\details{
Dane przeznaczone do wykorzystania z funkcjami pakietu IDEChoiceR,
w szczególności metodami TOPSIS oraz VIKOR.
}
