% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_topsis_wynik}
\alias{plot.rozmyty_topsis_wynik}
\title{Mapa Efektywności TOPSIS}
\usage{
\method{plot}{rozmyty_topsis_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_topsis_wynik`.}

\item{...}{Dodatkowe argumenty.
#' @export}
}
\description{
Pokazuje odległość od ideału. Oś X: Dystans od Najgorszego (D-).
Oś Y: Dystans do Najlepszego (D+).
Cel: Chcemy być w prawym dolnym rogu (Daleko od D-, Blisko D+).
}
