% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{prepare_ide_data}
\alias{prepare_ide_data}
\title{Przygotowanie macierzy decyzyjnej IDE}
\usage{
prepare_ide_data(
  data,
  model_syntax,
  alternative_column = NULL,
  aggregation_function = mean
)
}
\arguments{
\item{data}{Ramka danych zawierająca surowe oceny IDE.}

\item{model_syntax}{Definicja obszarów decyzyjnych w formie tekstowej,
np. "Wydajnosc =~ start + ram; Koszt =~ licencja + subskrypcja".}

\item{alternative_column}{Nazwa kolumny identyfikującej IDE.
Jeśli NULL, każdy wiersz traktowany jest jako osobna alternatywa.}

\item{aggregation_function}{Funkcja agregująca opinie ekspertów
(domyślnie: mean).}
}
\value{
Rozmyta macierz decyzyjna (m x 3n),
gdzie m oznacza liczbę środowisk IDE,
a n liczbę obszarów decyzyjnych.
}
\description{
Funkcja przekształca surowe dane dotyczące środowisk IDE
w ustrukturyzowaną (opcjonalnie rozmytą) macierz decyzyjną
gotową do analizy metodami TOPSIS lub VIKOR.

Oblicza wskaźniki kompozytowe dla zdefiniowanych obszarów
(np. Wydajność, Funkcjonalność, Użyteczność, Koszt),
dokonuje normalizacji do skali 1–9,
agreguje oceny ekspertów (jeśli występują)
oraz przeprowadza fuzzification.
}
